+++ src/test_QGLM.cpp
@@ -8,25 +8,25 @@
 #include <vector>
 
 // [[Rcpp::export]]
-Rcpp::List test_GLM(const std::string method,
-                    const Eigen::Map<Eigen::MatrixXd> &x,
-                    const Eigen::Map<Eigen::VectorXd> &par0,
-                    const bool intercept,
-                    const int maxit,
-                    const int maxit_l,
-                    const double tol,
-                    const double tol_l,
-                    const Rcpp::Nullable<double> step,
-                    const Rcpp::Nullable<double> th,
-                    const int nthreads,
-                    const Eigen::Map<Eigen::ArrayXd> &w)
+Rcpp::List test_QGLM(const std::string method,
+                     const Eigen::Map<Eigen::MatrixXd> &x,
+                     const Eigen::Map<Eigen::VectorXd> &par0,
+                     const bool intercept,
+                     const int maxit,
+                     const int maxit_l,
+                     const double tol,
+                     const double tol_l,
+                     const Rcpp::Nullable<double> step,
+                     const Rcpp::Nullable<double> th,
+                     const int nthreads,
+                     const Eigen::Map<Eigen::ArrayXd> &w)
 {
   const int p = x.cols() - 1;
   const double gamma = set_step(x.rows(), step);
 
   // overall test
-  Eigen::VectorXd par(p);
-  Eigen::VectorXd l(p);
+  Eigen::VectorXd par(p + 1);
+  Eigen::VectorXd l(p + 1);
   double nllr{};
   int iter{};
   bool conv{};
@@ -34,9 +34,10 @@ Rcpp::List test_GLM(const std::string method,
   double logl{};
   if (intercept && p > 1)
   {
-    Eigen::MatrixXd lhs(p - 1, p);
+    Eigen::MatrixXd lhs(p - 1, p + 1);
     lhs.col(0) = Eigen::MatrixXd::Zero(p - 1, 1);
-    lhs.rightCols(p - 1) = Eigen::MatrixXd::Identity(p - 1, p - 1);
+    lhs.middleCols(1, p - 1) = Eigen::MatrixXd::Identity(p - 1, p - 1);
+    lhs.col(p) = Eigen::MatrixXd::Zero(p - 1, 1);
     const Eigen::VectorXd rhs = Eigen::VectorXd::Zero(p - 1);
     const double test_th = set_threshold(p - 1, th);
     const CEL el(method, par0, x, lhs, rhs, maxit, maxit_l, tol, tol_l, gamma,
@@ -51,15 +52,20 @@ Rcpp::List test_GLM(const std::string method,
   }
   else
   {
-    par = Eigen::VectorXd::Zero(p);
+    Eigen::MatrixXd lhs(p, p + 1);
+    lhs.leftCols(p) = Eigen::MatrixXd::Identity(p, p);
+    lhs.col(p) = Eigen::MatrixXd::Zero(p, 1);
+    const Eigen::VectorXd rhs = Eigen::VectorXd::Zero(p);
     const double test_th = set_threshold(p, th);
-    const EL el(method, par, x, maxit_l, tol_l, test_th, w);
+    const CEL el(method, par0, x, lhs, rhs, maxit, maxit_l, tol, tol_l, gamma,
+                 test_th, w);
+    par = el.par;
     l = el.l;
     nllr = el.nllr;
     iter = el.iter;
     conv = el.conv;
-    logp = el.logp(x);
-    logl = el.loglik();
+    logp = el.logp(x, w);
+    logl = el.loglik(w);
   }
 
   // significance tests
@@ -72,7 +78,7 @@ Rcpp::List test_GLM(const std::string method,
   #endif
   for (int i = 0; i < p; ++i)
   {
-    Eigen::MatrixXd lhs = Eigen::MatrixXd::Zero(1, p);
+    Eigen::MatrixXd lhs = Eigen::MatrixXd::Zero(1, p + 1);
     lhs(i) = 1.0;
     const CEL par_test(method, par0, x, lhs, Eigen::VectorXd::Zero(1), maxit,
                        maxit_l, tol, tol_l, gamma, test_th, w);